feat: implement C++ feature parity — damage system, building lifecycle, GameObjManager, GameContext
Port of missing gameplay systems from C++ Renegade server:

- **GameContext**: session-scoped container replacing C++ global singletons
  (The_Game(), BaseControllerClass::CurrentBases, TimeManager::Get_Frame_Seconds)
- **GameObjManager**: drives think() loop for all game objects; owns building and star lists
- **DamageableGameObj**: add applyDamage() (shield-first model), completelyDamaged() hook,
  getNormalizedHealth/setNormalizedHealth, healthMax/shieldStrengthMax
- **BuildingGameObj**: mutable isDestroyed/isPowerOn/currentState, full state machine
  (updateState, percentageToHealthState, composeState), applyDamage override,
  onDestroyed, enablePower, cncInitialize, resetToFull
- **BaseControllerClass**: mutable fields with dirty-bit setters; checkBasePower,
  powerBase, checkRadar, onBuildingDestroyed, distributeFundsToEachTeammate,
  destroyBase, reset — canGenerateSoldiers/Vehicles default false (set by cncInitialize)
- **Building subtypes**: SoldierFactory/VehicleFactory/WarFactory cncInitialize +
  onDestroyed; PowerPlant triggers checkBasePower; ComCenter triggers checkRadar;
  Refinery passive credit trickle via think()
- **VehicleFactoryGameObj**: timer-driven requestVehicle/onGenerationComplete
- **GameState**: gameDurationSeconds Int→Float (matches C++ float GameDurationS)
- **ServerFps**: mutable fps via setFps() + BIT_FREQUENT dirty bit; dynamic fps measurement
- **SpawnManager**: return Pair<Vector3,Float> (position + facing)
- **God**: star list management (addStar/removeStar); playerData binding on soldier creation
- **BackgroundMgr/WeatherMgr**: creationDirtyBit=BIT_RARE (classId=0 must not use BIT_CREATION)
- **SoldierGameObj**: add playerData: Player? for fund distribution
- **GameServer**: integrate GameContext + GameObjManager; game loop reorder
  (gameObjManager.think → god.think → endGameTest → replication → deletePending);
  CHANGETEAMEVENT deletes soldier for respawn with new team; handleCoreRestart resets
  buildings + base controllers; register BackgroundMgr/WeatherMgr singletons;
  dynamic FPS measurement (updateFps); building cncInitialize during startup

+++ kotlin-server/server/src/main/kotlin/ccr/server/GameState.kt
@@ -13,7 +13,7 @@ class GameState(private val config: ServerConfig) {
         private set
 
     // gameDurationSeconds tracks elapsed play time (used for MVP qualifying time).
-    var gameDurationSeconds: Int = 0
+    var gameDurationSeconds: Float = 0f
         private set
 
     var frameCount: Int = 0
@@ -47,7 +47,7 @@ class GameState(private val config: ServerConfig) {
         }
 
         // Count elapsed play time
-        gameDurationSeconds += deltaSec.toInt()
+        gameDurationSeconds += deltaSec
 
         // Count down the timer (only if a time limit is set)
         if (timeRemainingSeconds > 0f) {
@@ -92,7 +92,7 @@ class GameState(private val config: ServerConfig) {
     /** Resets the game state for a new match (called at core restart). */
     fun reset() {
         timeRemainingSeconds = if (config.timeLimitMinutes > 0) config.timeLimitMinutes * 60f else 0f
-        gameDurationSeconds = 0
+        gameDurationSeconds = 0f
         frameCount = 0
         isIntermission = false
         intermissionTimeRemaining = 0f
